requirements.txt
appdirs==1.4.3
Django==1.11.1
django-chartit==0.2.9
django-crispy-forms==1.6.1
django-leaflet==0.22.0
django-registration==2.2
packaging==16.8
pkg-resources==0.0.0
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
